# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ClinicalCompass API",
    "summary": "Authentication API with database-backed opaque tokens.",
    "description": "\n**ClinicalCompass API** \u2014 username/password authentication using opaque,\nserver-side tokens (no JWT).\n\n### Auth flow\n1. `POST /auth/register` to create a user.\n2. `POST /auth/login` with the username + password to receive an opaque\n   `access_token`.\n3. Send the token as `Authorization: Bearer <access_token>` on protected\n   endpoints.\n4. `POST /auth/logout` to revoke the token.\n\nTokens are random strings stored in the database and expire after 24 hours.\n",
    "contact": {
      "name": "ClinicalCompass",
      "email": "support@example.com"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/submissions": {
      "post": {
        "tags": [
          "submissions"
        ],
        "summary": "Submit a module's form",
        "description": "Stores the authenticated user's answers for the given module's form. The answers are kept verbatim as a JSON object keyed by field name.",
        "operationId": "create_submission_submissions_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FormSubmissionCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormSubmissionOut"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Module not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "submissions"
        ],
        "summary": "List my form submissions (paginated)",
        "description": "Returns a page of the authenticated user's own submissions, newest first. Filter to one module with `module_id`.",
        "operationId": "list_submissions_submissions_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "module_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter to this module",
              "title": "Module Id"
            },
            "description": "Filter to this module"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Max items to return",
              "default": 20,
              "title": "Limit"
            },
            "description": "Max items to return"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of items to skip",
              "default": 0,
              "title": "Offset"
            },
            "description": "Number of items to skip"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormSubmissionPage"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/submissions/{submission_id}": {
      "get": {
        "tags": [
          "submissions"
        ],
        "summary": "Get one of my submissions",
        "operationId": "get_submission_submissions__submission_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "submission_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Submission Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormSubmissionOut"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Submission not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/submissions/{submission_id}/appeal-letter": {
      "get": {
        "tags": [
          "submissions"
        ],
        "summary": "Get the personalized appeal letter for a submission",
        "description": "Builds a personalized appeal letter from the submission's answers using the module's stored appeal-letter template (defaults, derived values, conditional sections, and {{placeholder}} substitution). Falls back to the module's plain-text appeal letter (with placeholder substitution) when no template is configured.",
        "operationId": "get_submission_appeal_letter_submissions__submission_id__appeal_letter_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "submission_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Submission Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppealLetterOut"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Submission not found, or no appeal letter configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AppealLetterOut": {
        "properties": {
          "submission_id": {
            "type": "integer",
            "title": "Submission Id"
          },
          "module_id": {
            "type": "integer",
            "title": "Module Id"
          },
          "letter": {
            "type": "string",
            "title": "Letter",
            "description": "The personalized appeal letter text."
          }
        },
        "type": "object",
        "required": [
          "submission_id",
          "module_id",
          "letter"
        ],
        "title": "AppealLetterOut",
        "description": "A rendered, personalized appeal letter for a submission.",
        "example": {
          "letter": "Dear UnitedHealthcare,\n\nRe: Jane Doe ...",
          "module_id": 1,
          "submission_id": 1
        }
      },
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "examples": [
              "Invalid username or password"
            ]
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "description": "Standard FastAPI error envelope."
      },
...
      "FormSubmissionCreate": {
        "properties": {
          "module_id": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Module Id",
            "description": "The module whose form is being submitted."
          },
          "data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Data",
            "description": "The submitted answers keyed by field name. Stored verbatim."
          }
        },
        "type": "object",
        "required": [
          "module_id"
        ],
        "title": "FormSubmissionCreate",
        "description": "A user's answers for a module's form.",
        "example": {
          "data": {
            "tStage": "T3",
            "tumorSize": 4.5
          },
          "module_id": 1
        }
      },
      "FormSubmissionOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "module_id": {
            "type": "integer",
            "title": "Module Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Data"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "module_id",
          "user_id",
          "data",
          "created_at",
          "updated_at"
        ],
        "title": "FormSubmissionOut"
      },
      "FormSubmissionPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/FormSubmissionOut"
            },
            "type": "array",
            "title": "Items"
          },
          "next": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next",
            "description": "URL of the next page, or null on the last page",
            "examples": [
              "http://localhost/submissions?limit=20&offset=20"
            ]
          },
          "previous": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous",
            "description": "URL of the previous page, or null on the first page",
            "examples": [
              null
            ]
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "FormSubmissionPage",
        "description": "A page of form submissions with next/previous page links."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}